+++ packages/backend/src/invoice/service.ts
@@ -1,5 +1,5 @@
 import { Invoice } from './model'
-import { AccountService } from '../account/service'
+import { AccountService, isAccountError } from '../account/service'
 import { BaseService } from '../shared/baseService'
 import assert from 'assert'
 import { Transaction } from 'knex'
@@ -61,10 +61,15 @@ async function createInvoice(
   const invTrx = trx || (await Invoice.startTransaction(deps.knex))
 
   try {
-    const subAccount = await deps.accountService.createSubAccount(
-      accountId,
+    const subAccount = await deps.accountService.create(
+      {
+        superAccountId: accountId
+      },
       invTrx
     )
+    if (isAccountError(subAccount)) {
+      throw new Error('unable to create account, err=' + subAccount)
+    }
 
     const invoice = await Invoice.query(invTrx).insertAndFetch({
       accountId,
