+++ packages/backend/src/shared/errors.ts
@@ -1,0 +1,33 @@
+import { AssetOptions } from '../asset/service'
+
+export class BalanceTransferError extends Error {
+  constructor(public code: number) {
+    super()
+    this.name = 'TransferError'
+  }
+}
+
+export class UnknownBalanceError extends Error {
+  constructor(accountId: string) {
+    super('Balance not found. accountId=' + accountId)
+    this.name = 'UnknownBalanceError'
+  }
+}
+
+export class UnknownLiquidityAccountError extends Error {
+  constructor(asset: AssetOptions) {
+    super(
+      'Unknown liquidity account. code=' + asset.code + ' scale=' + asset.scale
+    )
+    this.name = 'UnknownLiquidityAccountError'
+  }
+}
+
+export class UnknownSettlementAccountError extends Error {
+  constructor(asset: AssetOptions) {
+    super(
+      'Unknown settlement account. code=' + asset.code + ' scale=' + asset.scale
+    )
+    this.name = 'UnknownSettlementAccountError'
+  }
+}
