+++ packages/backend/src/webmonetization/service.test.ts
@@ -6,6 +6,7 @@ import { Config } from '../config/app'
 import { IocContract } from '@adonisjs/fold'
 import { initIocContainer } from '../'
 import { AppServices } from '../app'
+import { AccountFactory } from '../tests/accountFactory'
 import { truncateTables } from '../tests/tableManager'
 import { Account } from '../account/model'
 import { Invoice } from '../invoice/model'
@@ -15,6 +16,7 @@ describe('WM Service', (): void => {
   let deps: IocContract<AppServices>
   let appContainer: TestContainer
   let wmService: WebMonetizationService
+  let accountFactory: AccountFactory
   let knex: Knex
   let account: Account
   const mockMessageProducer = {
@@ -33,9 +35,8 @@ describe('WM Service', (): void => {
   beforeEach(
     async (): Promise<void> => {
       wmService = await deps.use('wmService')
-      account = await deps.use('accountService').then((as) => {
-        return as.create(6, 'USD')
-      })
+      accountFactory = new AccountFactory(await deps.use('accountService'))
+      account = await accountFactory.build()
     }
   )
